DM: lib: cleanup and interface change for safe2remove for FESA
