+++ modules/ftm/ftmx86/Makefile
@@ -1,6 +1,6 @@
 TOP          ?= ../../..
 
-VERSION      ?= 1.1.0
+VERSION      ?= 1.2.0
 VERSION_TOOL ?= 0.9.0
 
 
@@ -34,7 +34,7 @@ endif
 T_CXX         = g++
 T_CXXFLAGS    = -Wall -g -std=c++11 -I$(FTMPATH)/include $(EB_INC) -I$(FWPATH) -Wfatal-errors -DPLATFORM=$(BUILD) -DTOOL_VER=\"$(VERSION_TOOL)\"  -DEXP_VER=\"$(VERSION)\" -DETHERBONE_THROWS=1
 
-SRC_FILES     = event.cpp meta.cpp block.cpp visitoruploadcrawler.cpp visitordownloadcrawler.cpp visitorvertexwriter.cpp hashmap.cpp carpeDM.cpp carpeDMcommand.cpp carpeDMuploadschedule.cpp carpeDMdownloadschedule.cpp graph.cpp alloctable.cpp mempool.cpp dotstr.cpp idformat.cpp grouptable.cpp validation.cpp visitorvalidation.cpp common.cpp
+SRC_FILES     = event.cpp meta.cpp block.cpp visitoruploadcrawler.cpp visitordownloadcrawler.cpp visitorvertexwriter.cpp hashmap.cpp carpeDM.cpp carpeDMcommand.cpp carpeDMuploadschedule.cpp carpeDMdownloadschedule.cpp graph.cpp alloctable.cpp mempool.cpp dotstr.cpp idformat.cpp grouptable.cpp validation.cpp visitorvalidation.cpp common.cpp carpeDMsafe2remove.cpp
 T_SOURCES     = $(addprefix $(SRCPATH)/, $(SRC_FILES))
 T_OBJECTS     = $(subst .cpp,.o,$(SRC_FILES))
 T_LIBS        = -Wl,-rpath,/usr/local/lib $(EB_LIB) -lboost_graph $(SER_LIB)
