+++ syn/gsi_pexarria5/ftm/Makefile
@@ -5,7 +5,7 @@ SPI_LANES	= ASx4
 RAM_SIZE	= 131072 
 SHARED_SIZE     = 98304 
 USRCPUCLK       = 125000
-VERSION         = 1.1.0
+VERSION         = 1.2.0
 VERSION_TOOL    = 0.9.0
 DEBUGLVL  	= 0
